+++ src/models/groups/archives/dto/groupArchiveModel.dto.ts
@@ -11,6 +11,9 @@ export class GroupArchiveModelDto {
     @ApiProperty({ description: 'Id профиля записи', type: Number })
     profileId: number;
 
+    @ApiProperty({ description: 'Id профиля партнёра', type: Number, required: false })
+    partnerId?: number;
+
     @ApiProperty({ description: 'Дата внесения группы в архив', type: Date })
     createdAt: Date;
 }