+++ src/models/users/accounts/config/account.config.ts
@@ -1,5 +1,5 @@
 import {Prisma} from "@prisma/client";
-import {UserMinimalPConfig} from "@models/users/config";
+import {UserMinimalPConfig} from "@models/users/config/userMinimal.config";
 import {AccountModelPConfig} from "./accountModel.config";