+++ src/models/users/profiles/config/profile.config.ts
@@ -1,9 +1,10 @@
+import {GroupMinimalPConfig} from "@models/groups/config/groupMinimal.config";
 import {ProfileModelPConfig} from "./profileModel.config";
 import {Prisma} from "@prisma/client";
 
 
 export const ProfilePConfig: Prisma.ProfileSelect = {
     ...ProfileModelPConfig,
-    mainGroup: true,
-    secondGroup: true
+    mainGroup: { select: GroupMinimalPConfig },
+    secondGroup: { select: GroupMinimalPConfig }
 }