+++ src/models/users/profiles/dto/profile.dto.ts
@@ -1,9 +1,9 @@
 import {ApiProperty} from "@nestjs/swagger";
-import {GroupModelDto} from "@models/groups/dto/groupModel.dto";
 import {ProfileModelDto} from "./profileModel.dto";
+import {GroupMinimalDto} from "@models/groups/dto";
 
 
 export class ProfileDto extends ProfileModelDto {
-    @ApiProperty({ description: 'Активная группа', type: GroupModelDto, required: false })
-    group?: GroupModelDto;
+    @ApiProperty({ description: 'Активная группа', type: GroupMinimalDto, required: false })
+    group?: GroupMinimalDto;
 }