+++ src/modules/app/app.module.ts
@@ -4,11 +4,13 @@ import {Module} from '@nestjs/common';
 import {ConfigModule} from "@nestjs/config";
 import {ScheduleModule} from "@nestjs/schedule";
 import {SessionsModule} from "@modules/sessions/sessions.module";
+import {HdRezkaModule} from "@modules/hdRezka/hdRezka.module";
 import {PrismaModule} from "@modules/prisma/prisma.module";
 import {GroupsModule} from "@modules/groups/groups.module";
 import {MoviesModule} from "@modules/movies/movies.module";
 import {TasksModule} from "@modules/tasks/tasks.module";
 import {UsersModule} from "@modules/users/users.module";
+import {MailsModule} from "@modules/mails/mails.module";
 import {FilesModule} from "@modules/files/files.module";
 import {AuthModule} from "@modules/auth/auth.module";
 import {AppController} from "./app.controller";
@@ -25,6 +27,8 @@ import {AppService} from "./app.service";
             }
         ]),
         PrismaModule,
+        //HdRezkaModule,
+        //MailsModule,
         TasksModule,
         AuthModule,
         FilesModule,
@@ -38,10 +42,10 @@ import {AppService} from "./app.service";
     ],
     providers: [
         AppService,
-        {
-            provide: APP_GUARD,
-            useClass: ThrottlerGuard
-        }
+        // {
+        //     provide: APP_GUARD,
+        //     useClass: ThrottlerGuard
+        // }
     ]
 })
 export class AppModule {}