+++ src/modules/auth/strategy/accessToken.strategy.ts
@@ -8,6 +8,7 @@ import {SessionsModelService} from "@models/sessions/sessions.service";
 import {UsersModelService} from "@models/users/users.service";
 import {PayloadReturnDto, TokenPayloadDto} from "./dto";
 import validate from './validate';
+import {GroupsModelService} from "@models/groups/groups.service";
 
 @Injectable()
 export class AccessTokenStrategy extends PassportStrategy(Strategy, 'jwt') {
@@ -15,6 +16,7 @@ export class AccessTokenStrategy extends PassportStrategy(Strategy, 'jwt') {
         private configService: ConfigService,
         private sessionsModelService: SessionsModelService,
         private usersModelService: UsersModelService,
+        private groupsModelService: GroupsModelService,
         private usersProfilesModelService: UsersProfilesModelService
     ) {
         super({
@@ -25,6 +27,14 @@ export class AccessTokenStrategy extends PassportStrategy(Strategy, 'jwt') {
     }
 
     async validate(req: Request, payload: TokenPayloadDto): Promise<PayloadReturnDto> {
-        return await validate('access', req, payload, this.sessionsModelService, this.usersModelService, this.usersProfilesModelService);
+        return await validate(
+            'access',
+            req,
+            payload,
+            this.sessionsModelService,
+            this.usersModelService,
+            this.groupsModelService,
+            this.usersProfilesModelService
+        );
     }
 }