+++ src/modules/hdRezka/dto/hdrMoviePart.dto.ts
@@ -1,0 +1,10 @@
+import {MovieType} from "@prisma/client";
+
+export class HdrMoviePartDto {
+    name: string;
+    link: string;
+    current: boolean;
+    releaseYear: number;
+    type: MovieType;
+    rating: number | null;
+}