+++ src/modules/hdRezka/dto/hdrMovieSeason.dto.ts
@@ -1,0 +1,7 @@
+import {HdrMovieSeriaDto} from "./hdrMovieSeria.dto";
+
+export class HdrMovieSeasonDto {
+    number: number;
+    releaseDate: Date;
+    series: HdrMovieSeriaDto[];
+}