+++ src/modules/hdRezka/dto/hdrMovieSeria.dto.ts
@@ -1,0 +1,5 @@
+export class HdrMovieSeriaDto {
+    name: string;
+    number: number;
+    releaseDate: Date;
+}