+++ src/modules/hdRezka/dto/hdrSearchReq.dto.ts
@@ -1,0 +1,8 @@
+import {HdrSearchValueDto} from "./hdrSearchValue.dto";
+import {HdrReqStatusInterface} from "../interfaces";
+
+
+export type HdrSearchReq = {
+    status: HdrReqStatusInterface;
+    values: HdrSearchValueDto[];
+}