+++ src/modules/hdRezka/dto/hdrSearchValue.dto.ts
@@ -1,0 +1,9 @@
+import {MovieType} from "@prisma/client";
+
+export type HdrSearchValueDto = {
+    name: string;
+    addName: string;
+    url: string;
+    type: MovieType;
+    rating: number | null;
+}