+++ src/modules/hdRezka/interfaces/hdrReqStatus.interface.ts
@@ -1,0 +1,5 @@
+export enum HdrReqStatusInterface {
+    OK = 'OK',
+    ERROR = 'ERROR',
+    NO_MIRROR = 'NO_MIRROR'
+}