+++ src/modules/tasks/queues/xmlParser.queue.ts
@@ -8,6 +8,7 @@ import {FilesService} from "@modules/files/files.service";
 import {Movie} from "@prisma/client";
 import { utils } from "@root/helpers";
 
+
 @Injectable()
 @SkipThrottle()
 export class XmlParserQueue {